requirements.txt
django<1.10
tri.table